Move link test utilities into single test using them
Also removes link test utilities from the public API.
