+++ src/link/core/__tests__/link.ts
@@ -3,7 +3,6 @@ import gql from 'graphql-tag';
 import { print } from 'graphql/language/printer';
 
 import { execute, ApolloLink, from, split, concat } from '../link';
-import { testLinkResults } from '../test-utils';
 import { FetchResult, Operation, NextLink, GraphQLRequest } from '../types';
 
 class SetContextLink extends ApolloLink {
@@ -31,6 +30,45 @@ const sampleQuery = gql`
     }
   }
 `;
+
+function checkCalls<T>(calls: any[] = [], results: Array<T>) {
+  expect(calls.length).toBe(results.length);
+  calls.map((call, i) => expect(call.data).toEqual(results[i]));
+}
+
+interface TestResultType {
+  link: ApolloLink;
+  results?: any[];
+  query?: string;
+  done?: () => void;
+  context?: any;
+  variables?: any;
+}
+
+function testLinkResults(params: TestResultType) {
+  const { link, context, variables } = params;
+  const results = params.results || [];
+  const query = params.query || sampleQuery;
+  const done = params.done || (() => void 0);
+
+  const spy = jest.fn();
+  execute(link, { query, context, variables }).subscribe({
+    next: spy,
+    error: error => {
+      expect(error).toEqual(results.pop());
+      checkCalls(spy.mock.calls[0], results);
+      if (done) {
+        done();
+      }
+    },
+    complete: () => {
+      checkCalls(spy.mock.calls[0], results);
+      if (done) {
+        done();
+      }
+    },
+  });
+}
 
 const setContext = () => ({ add: 1 });
 describe('ApolloLink(abstract class)', () => {
