--- src/link/core/test-utils.ts
@@ -1,1 +1,0 @@
-export * from './test-utils/testingUtils';
